refactor: put PreviewSerializer etc in Cases module
