+++ app/serializers/cases/preview_serializer.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Cases
+  # A subset of case data
+  class PreviewSerializer < ApplicationSerializer
+    attributes :acknowledgements, :audience, :authors, :commentable, :cover_url,
+               :dek, :featured_at, :kicker, :latitude, :learning_objectives,
+               :longitude, :other_available_locales, :photo_credit,
+               :published_at, :slug, :small_cover_url, :summary, :title,
+               :translators, :zoom
+
+    link(:self) { case_path(I18n.locale, object) }
+    link(:settings) { edit_case_settings_path(I18n.locale, object) }
+    link(:teach) { new_deployment_path(case_slug: object.slug) }
+  end
+end
